Move tests as a template test
