+++ Sofa/framework/LinearAlgebra/Testing/CMakeLists.txt
@@ -4,6 +4,7 @@ project(Sofa.LinearAlgebra.Testing LANGUAGES CXX)
 set(HEADER_FILES
     src/Sofa.LinearAlgebra.Testing/BaseMatrix_test.h
     src/Sofa.LinearAlgebra.Testing/SparseMatrixTest.h
+    src/Sofa.LinearAlgebra.Testing/SparseMatrixProduct_test.h
 )
 
 add_library(${PROJECT_NAME} INTERFACE)
